some changes in Table
