+++ client/src/components/General/Table/BookingDetail/index.js
@@ -6,203 +6,203 @@ import classNames from 'classnames/bind';
 
 
 const cx = classNames.bind(styles);
-function BookingDetail() {
-    const columns = [
-        {
-            title: 'Mã khách sạn',
-            dataIndex: 'name',
-            key: 'name',
-            render: (text) => <a>{text}</a>,
-        },
-        {
-            title: 'Khách sạn',
-            dataIndex: 'age',
-            key: 'age',
-        },
-        {
-            title: 'Tên khách hàng',
-            dataIndex: 'address',
-            key: 'address',
-        },
-        {
-            title: 'Ngày nhận phòng',
-            key: 'tags',
-            dataIndex: 'tags',
-            render: (_, { tags }) => (
-                <>
-                    {tags.map((tag) => {
-                        let color = tag.length > 5 ? 'geekblue' : 'green';
-                        if (tag === 'loser') {
-                            color = 'volcano';
-                        }
-                        return (
-                            <Tag color={color} key={tag}>
-                                {tag.toUpperCase()}
-                            </Tag>
-                        );
-                    })}
-                </>
-            ),
-        },
-        {
-            title: 'Ngày trả phòng',
-            key: 'action',
-            render: (_, record) => (
-                <Space size="middle">
-                    <a>Invite {record.name}</a>
-                    <a>Delete</a>
-                </Space>
-            ),
-        },
-        {
-            title: 'Trạng thái',
-            key: 'tags',
-            dataIndex: 'tags',
-            render: (_, { tags }) => (
-                <>
-                    {tags.map((tag) => {
-                        let color = tag.length > 5 ? 'geekblue' : 'green';
-                        if (tag === 'loser') {
-                            color = 'volcano';
-                        }
-                        return (
-                            <Tag color={color} key={tag}>
-                                {tag.toUpperCase()}
-                            </Tag>
-                        );
-                    })}
-                </>
-            ),
-        },
-    ];
-    const data = [
-        {
-            key: '1',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '2',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '3',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-        {
-            key: '4',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '5',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '6',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-        {
-            key: '7',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '8',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '9',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-        {
-            key: '10',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '11',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '12',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-        {
-            key: '13',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '14',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '15',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-        {
-            key: '16',
-            name: 'John Brown',
-            age: 32,
-            address: 'New York No. 1 Lake Park',
-            tags: ['nice', 'developer'],
-        },
-        {
-            key: '17',
-            name: 'Jim Green',
-            age: 42,
-            address: 'London No. 1 Lake Park',
-            tags: ['loser'],
-        },
-        {
-            key: '18',
-            name: 'Joe Black',
-            age: 32,
-            address: 'Sidney No. 1 Lake Park',
-            tags: ['cool', 'teacher'],
-        },
-    ];
+function BookingDetail({columns, data}) {
+    // const columns = [
+    //     {
+    //         title: 'Mã khách sạn',
+    //         dataIndex: 'name',
+    //         key: 'name',
+    //         render: (text) => <a>{text}</a>,
+    //     },
+    //     {
+    //         title: 'Khách sạn',
+    //         dataIndex: 'age',
+    //         key: 'age',
+    //     },
+    //     {
+    //         title: 'Tên khách hàng',
+    //         dataIndex: 'address',
+    //         key: 'address',
+    //     },
+    //     {
+    //         title: 'Ngày nhận phòng',
+    //         key: 'tags',
+    //         dataIndex: 'tags',
+    //         render: (_, { tags }) => (
+    //             <>
+    //                 {tags.map((tag) => {
+    //                     let color = tag.length > 5 ? 'geekblue' : 'green';
+    //                     if (tag === 'loser') {
+    //                         color = 'volcano';
+    //                     }
+    //                     return (
+    //                         <Tag color={color} key={tag}>
+    //                             {tag.toUpperCase()}
+    //                         </Tag>
+    //                     );
+    //                 })}
+    //             </>
+    //         ),
+    //     },
+    //     {
+    //         title: 'Ngày trả phòng',
+    //         key: 'action',
+    //         render: (_, record) => (
+    //             <Space size="middle">
+    //                 <a>Invite {record.name}</a>
+    //                 <a>Delete</a>
+    //             </Space>
+    //         ),
+    //     },
+    //     {
+    //         title: 'Trạng thái',
+    //         key: 'tags',
+    //         dataIndex: 'tags',
+    //         render: (_, { tags }) => (
+    //             <>
+    //                 {tags.map((tag) => {
+    //                     let color = tag.length > 5 ? 'geekblue' : 'green';
+    //                     if (tag === 'loser') {
+    //                         color = 'volcano';
+    //                     }
+    //                     return (
+    //                         <Tag color={color} key={tag}>
+    //                             {tag.toUpperCase()}
+    //                         </Tag>
+    //                     );
+    //                 })}
+    //             </>
+    //         ),
+    //     },
+    // ];
+    // const data = [
+    //     {
+    //         key: '1',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '2',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '3',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    //     {
+    //         key: '4',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '5',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '6',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    //     {
+    //         key: '7',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '8',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '9',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    //     {
+    //         key: '10',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '11',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '12',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    //     {
+    //         key: '13',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '14',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '15',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    //     {
+    //         key: '16',
+    //         name: 'John Brown',
+    //         age: 32,
+    //         address: 'New York No. 1 Lake Park',
+    //         tags: ['nice', 'developer'],
+    //     },
+    //     {
+    //         key: '17',
+    //         name: 'Jim Green',
+    //         age: 42,
+    //         address: 'London No. 1 Lake Park',
+    //         tags: ['loser'],
+    //     },
+    //     {
+    //         key: '18',
+    //         name: 'Joe Black',
+    //         age: 32,
+    //         address: 'Sidney No. 1 Lake Park',
+    //         tags: ['cool', 'teacher'],
+    //     },
+    // ];
     return (
         <Table classNames={cx("table")} columns={columns} dataSource={data}></Table>
     )
